requirements.txt
Django
django-modeltranslation
django-meta
django-tinymce
django-simple-open-graph
django-rest-framework
django-filter